# backend/swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/onboarding": {
      "post": {
        "operationId": "OnboardingController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOnboardingDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Onboarding process created successfully."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new onboarding process",
        "tags": [
          "Onboarding"
        ]
      },
      "get": {
        "operationId": "OnboardingController_findAll",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "ACTIVE",
                "COMPLETED",
                "EXTENDED"
              ],
              "type": "string"
            }
          },
          {
            "name": "mentorId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Onboardings list retrieved successfully."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all onboarding processes with optional status/mentor filters",
        "tags": [
          "Onboarding"
        ]
      }
    },
    "/onboarding/dashboard/stats": {
      "get": {
        "operationId": "OnboardingController_getMilestoneTrackingStats",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Dashboard stats retrieved."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get milestone tracking statistics for dashboard",
        "tags": [
          "Onboarding"
        ]
      }
    },
    "/onboarding/dashboard/at-risk": {
      "get": {
        "operationId": "OnboardingController_getAtRiskDevelopers",
        "parameters": [],
        "responses": {
          "200": {
            "description": "At-risk developers list retrieved."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get list of developers at risk during onboarding",
        "tags": [
          "Onboarding"
        ]
      }
    },
    "/onboarding/stats": {
      "get": {
        "operationId": "OnboardingController_getDashboardStats",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Stats retrieved."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get general onboarding dashboard stats",
        "tags": [
          "Onboarding"
        ]
      }
    },
    "/onboarding/velocity": {
      "get": {
        "operationId": "OnboardingController_getVelocityComparison",
        "parameters": [
          {
            "name": "teamId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Velocity comparison stats retrieved."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get onboarding velocity comparison across teams",
        "tags": [
          "Onboarding"
        ]
      }
    },
    "/onboarding/velocity/{userId}": {
      "get": {
        "operationId": "OnboardingController_getDeveloperVelocity",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "description": "Developer User ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Developer velocity stats retrieved."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get onboarding velocity metrics for an individual developer",
        "tags": [
          "Onboarding"
        ]
      }
    },
    "/onboarding/templates": {
      "get": {
        "operationId": "OnboardingController_getTemplates",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Templates list retrieved."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get onboarding checklists templates",
        "tags": [
          "Onboarding"
        ]
      },
      "post": {
        "operationId": "OnboardingController_createTemplate",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTemplateDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Checklist template created successfully."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new onboarding checklist template",
        "tags": [
          "Onboarding"
        ]
      }
    },
    "/onboarding/templates/{id}": {
      "get": {
        "operationId": "OnboardingController_getTemplate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Template ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Template details retrieved."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get onboarding template details by ID",
        "tags": [
          "Onboarding"
        ]
      },
      "patch": {
        "operationId": "OnboardingController_updateTemplate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Template ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTemplateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Template updated successfully."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update template details by ID",
        "tags": [
          "Onboarding"
        ]
      },
      "delete": {
        "operationId": "OnboardingController_deleteTemplate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Template ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Template successfully deleted."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a template by ID",
        "tags": [
          "Onboarding"
        ]
      }
    },
    "/onboarding/{userId}/progress": {
      "get": {
        "operationId": "OnboardingController_getDeveloperProgress",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "description": "Developer User ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Developer progress data retrieved."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get individual developer progress metrics",
        "tags": [
          "Onboarding"
        ]
      }
    },
    "/onboarding/{userId}/timeline": {
      "get": {
        "operationId": "OnboardingController_getMilestoneTimeline",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "description": "Developer User ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Milestone timeline entries retrieved."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get individual milestone timeline",
        "tags": [
          "Onboarding"
        ]
      }
    },
    "/onboarding/mentors/available": {
      "get": {
        "operationId": "OnboardingController_getAvailableMentors",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Available mentors list retrieved."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get list of available mentors for assignment",
        "tags": [
          "Onboarding"
        ]
      }
    },
    "/onboarding/mentors/{mentorId}/capacity": {
      "get": {
        "operationId": "OnboardingController_getMentorCapacity",
        "parameters": [
          {
            "name": "mentorId",
            "required": true,
            "in": "path",
            "description": "Mentor User ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Mentor capacity information retrieved."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get capacity status for a mentor",
        "tags": [
          "Onboarding"
        ]
      }
    },
    "/onboarding/{id}": {
      "get": {
        "operationId": "OnboardingController_findById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Onboarding ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Onboarding details retrieved."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get onboarding process details by ID",
        "tags": [
          "Onboarding"
        ]
      }
    },
    "/onboarding/{id}/mentor": {
      "patch": {
        "operationId": "OnboardingController_assignMentor",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Onboarding ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Mentor successfully assigned."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Assign a mentor to an onboarding process",
        "tags": [
          "Onboarding"
        ]
      }
    },
    "/onboarding/{id}/extend": {
      "patch": {
        "operationId": "OnboardingController_extend",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Onboarding ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExtendOnboardingDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Onboarding extended successfully."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Extend onboarding duration",
        "tags": [
          "Onboarding"
        ]
      }
    },
    "/onboarding/{id}/complete": {
      "patch": {
        "operationId": "OnboardingController_complete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Onboarding ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Onboarding marked as complete."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mark onboarding process as complete",
        "tags": [
          "Onboarding"
        ]
      }
    },
    "/onboarding/{id}/checklist": {
      "get": {
        "operationId": "OnboardingController_getChecklist",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Onboarding ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Checklist items status list retrieved."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get onboarding checklist items status",
        "tags": [
          "Onboarding"
        ]
      }
    },
    "/onboarding/{id}/checklist/{itemId}": {
      "patch": {
        "operationId": "OnboardingController_updateChecklistItem",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Onboarding ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "itemId",
            "required": true,
            "in": "path",
            "description": "Checklist Item ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateChecklistItemDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Checklist item updated successfully."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update status of a specific checklist item",
        "tags": [
          "Onboarding"
        ]
      }
    },
...
  },
  "info": {
    "title": "My API",
    "description": "API documentation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "CreateOnboardingDto": {
        "type": "object",
        "properties": {}
      },
      "CreateTemplateDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateTemplateDto": {
        "type": "object",
        "properties": {}
      },
      "ExtendOnboardingDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateChecklistItemDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}